refactor: separate controls into different files
